Rename the data/ocean_emulators package to ocean_preprocessing to end the import/distribution-name collision with the ML modeling package at the repo root. The two subprojects are now fully decoupled: the modeling side stays on uv, the preprocessing side stays on conda/mamba (unavoidable because of xESMF), and neither imports the other. Tooling is unified under a single root pre-commit config, with separate ruff hooks per subproject so each uses its own config.
