+++ .pre-commit-config.yaml
@@ -4,9 +4,21 @@ repos:
   hooks:
     - id: ruff
       args: ["--fix", "--config", "pyproject.toml"]
-      exclude: ^notebooks/
+      exclude: ^(notebooks/|data/)
     - id: ruff-format
-      exclude: ^notebooks/
+      exclude: ^(notebooks/|data/)
+    - id: ruff
+      alias: ruff-data
+      name: ruff (data/)
+      args: ["--fix", "--config", "data/pyproject.toml"]
+      files: ^data/
+      exclude: ^data/notebooks/
+    - id: ruff-format
+      alias: ruff-format-data
+      name: ruff-format (data/)
+      args: ["--config", "data/pyproject.toml"]
+      files: ^data/
+      exclude: ^data/notebooks/
 - repo: https://github.com/pre-commit/pre-commit-hooks
   rev: v5.0.0
   hooks:
@@ -30,7 +42,7 @@ repos:
     entry: uv run --dev mypy
     types_or: [python, pyi]
     args: ["--config", "pyproject.toml"]
-    exclude: ^notebooks/,^data/notebooks/
+    exclude: ^(notebooks/|data/)
 - repo: https://github.com/tox-dev/pyproject-fmt
   rev: "v2.5.0"
   hooks:
