--- data/ocean_emulators/__init__.py
@@ -1,4 +1,0 @@
-try:
-    from ._version import __version__
-except ImportError:
-    __version__ = "unknown"
