+++ data/tests/test_data.py
@@ -1,7 +1,7 @@
 """This is s bit silly. It would be better to be able to create datasets from the schema dynamically, so the schema serve as source of truth, and when they are updated, the test data is too. But for now lets at least test that our test fixtures are compliant."""
 
 from tests.data import input_data, prediction_data, processed_data  # noqa
-from ocean_emulators.dataset_validation import (
+from ocean_preprocessing.dataset_validation import (
     ds_processed_validate,
     ds_input_validate,
     ds_prediction_validate,
