+++ data/tests/test_utils.py
@@ -1,8 +1,12 @@
 import numpy as np
 import pytest
 import xarray as xr
+from ocean_preprocessing.utils import (
+    apply_mask,
+    assert_mask_match,
+    ensure_nan_consistency,
+)
 
-from ocean_emulators.utils import apply_mask, assert_mask_match, ensure_nan_consistency
 from tests.data import (
     input_data,  # noqa # Might want to put these in conftest.py (see https://stackoverflow.com/questions/73191533/using-conftest-py-vs-importing-fixtures-from-dedicate-modules)
 )
