+++ notebooks/check_zarr.py
@@ -1,11 +1,9 @@
 #!/usr/bin/env python
-# coding: utf-8
 
 # In[1]:
 
 
 import xarray as xr
-
 
 # In[2]:
