+++ notebooks/cloud_data_check.py
@@ -1,5 +1,4 @@
 #!/usr/bin/env python
-# coding: utf-8
 
 # In[1]:
 
@@ -91,7 +90,6 @@ get_ipython().run_cell_magic(
 
 
 import xarray as xr
-
 
 # In[10]:
